+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 
 ## Unreleased
 
+### Add
+
+* Use an existing, immutable `Transaction` as a template for a new one via `TransactionBuilder.cloneFrom(tx)` ([TODO]()).
+
 
 ## [v9.0.0](https://github.com/stellar/js-stellar-base/compare/v8.2.2..v9.0.0)